pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.5.0",
    "jinaai>=0.2.10",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
]
